add config reboot

+++ disable_selinux_reboot.sh
@@ -3,4 +3,12 @@
 echo "Disable selinux"
 
 sudo sed -i s/^SELINUX=.*$/SELINUX=disabled/ /etc/selinux/config
-sudo reboot
+
+read -r -p "Are you sure? [y/N] " response
+if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]
+then
+    sudo reboot
+else
+    echo 'do not reboot'
+fi
+
